# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Health Hub API
  version: 0.20.0
  description: |
    API для приложения "Центр здоровья".
    Canonical file — все эндпоинты описаны здесь.

    v0.20.0: Added Food Preferences API (GET/POST/DELETE /v1/food/prefs) and Meal Plans API (GET/PUT/DELETE /v1/meal/plan, GET /v1/meal/today). Extended FeedDayResponse with meal_today, meal_plan_title, food_prefs_count.
    v0.19.0: Added Nutrition Targets API endpoints GET/PUT /v1/nutrition/targets and extended FeedDayResponse with nutrition_targets and nutrition_progress.
    v0.18.0: Added Workout Plans API endpoints and nutrition_plan proposal support.
    v0.17.0: Added supplement schedules API and vitamins_schedule proposal apply flow.
    v0.16.0: Added AI proposal management endpoints GET /v1/ai/proposals, POST /v1/ai/proposals/{id}/apply, POST /v1/ai/proposals/{id}/reject.
    v0.15.0: Added Chat MVP endpoint GET/POST /v1/chat/messages with assistant proposals.
    v0.14.0: Added user settings endpoints GET/PUT /v1/settings and notifications settings persistence.
    v0.13.0: Added Email OTP auth endpoints POST /v1/auth/email/request and POST /v1/auth/email/verify.
    v0.12.0: Added POST /v1/auth/siwa with Apple identity token verification via JWKS.
    v0.11.0: Added POST /v1/auth/dev for local dev JWT issuance.
    v0.10.0: Added Sign in with Apple (SIWA) + JWT authentication, Intakes (Water & Supplements) API.
    v0.9.0: Added Inbox/Notifications API for server-side notifications.
...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /v1/sources:
    post:
      summary: Create source (link/note)
      description: Создание link или note source
      operationId: createSource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSourceRequest"
      responses:
        "201":
          description: Source создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceDTO"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    get:
      summary: List sources
      description: Получение списка sources для профиля с опциональной фильтрацией
      operationId: listSources
      parameters:
        - name: profile_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: checkin_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: query
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Список sources
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourcesResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v1/sources/{id}:
    delete:
      summary: Delete source
      description: Удаление source и связанных данных (S3 объекта при наличии)
      operationId: deleteSource
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Source удалён
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # === Inbox/Notifications API ===
...
components:
  schemas:
...
    CreateSourceRequest:
      type: object
      properties:
        profile_id:
          type: string
          format: uuid
        kind:
          type: string
          enum: [link, note]
        title:
          type: string
        text:
          type: string
          description: "Обязательно для kind=note"
        url:
          type: string
          description: "Обязательно для kind=link"
        checkin_id:
          type: string
          format: uuid
      required: [profile_id, kind]
    SourceDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        profile_id:
          type: string
          format: uuid
        kind:
          type: string
          enum: [link, note, image]
        title:
          type: string
        text:
          type: string
        url:
          type: string
        checkin_id:
          type: string
          format: uuid
        content_type:
          type: string
          description: "MIME type для images"
        size_bytes:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
      required: [id, profile_id, kind, created_at]
    SourcesResponse:
      type: object
      properties:
        sources:
          type: array
          items:
            $ref: "#/components/schemas/SourceDTO"
      required: [sources]

    # --- Notifications/Inbox ---
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Код ошибки
              example: "invalid_range"
            message:
              type: string
              description: Описание ошибки
              example: "Date range exceeds maximum of 90 days"
          required: [code, message]
      required: [error]
  responses:
    BadRequest:
      description: Неверный запрос
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Внутренняя ошибка сервера
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT токен, полученный через POST /v1/auth/dev, /v1/auth/siwa или /v1/auth/email/verify.
        При AUTH_MODE=none токен не обязателен; при наличии токена сервер валидирует его.
        Header: Authorization: Bearer <jwt>

# По умолчанию аутентификация НЕ применяется (AUTH_MODE=none).
# Когда включится строгая защита, добавить security на защищённые эндпоинты:
# security:
#   - BearerAuth: []